+++ .tape.js
@@ -1,14 +1,17 @@
 module.exports = {
 	basic: {
 		message: 'supports basic usage'
+	},
+	remove: {
+		message: 'supports remove usage'
+	},
+	replace: {
+		message: 'supports replace usage'
+	},
+	ignore: {
+		message: 'supports ignore usage'
+	},
+	tags: {
+		message: 'supports tags usage'
 	}
-	// remove: {
-	// 	message: 'supports remove usage'
-	// },
-	// replace: {
-	// 	message: 'supports replace usage'
-	// },
-	// tags: {
-	// 	message: 'supports tags usage'
-	// }
 };