+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changes to PostCSS Editor Styles
 
+### 0.2.0 (October 7, 2019)
+
+- properly handle options as arrays.
+- only change the tag if it's at the end.
+- add some more editor specific tags to defaults.
+
+
 ### 0.1.0 (October 4, 2018)
 
 - Initial version